Revert "moved .py to fs42 and update .sh"
This reverts commit b5f521fb61e720a10845b93429ac521294f685ef.
