+++ hot_start.sh
@@ -4,5 +4,5 @@
 cd $(dirname $0)
 . env/bin/activate
 
-python3 fs42/field_player.py 1>/dev/null 2>/dev/null & disown
+python3 field_player.py 1>/dev/null 2>/dev/null & disown
 python3 fs42/command_input.py 1>/dev/null 2>/dev/null & disown
