+++ rebuild_catalog.sh
@@ -4,4 +4,4 @@
 cd $(dirname $0)
 . env/bin/activate
 
-python3 fs42/station_42.py --rebuild_catalog
+python3 station_42.py --rebuild_catalog
