+++ test.sh
@@ -4,4 +4,4 @@
 cd $(dirname $0)
 . env/bin/activate
 
-python3 fs42/field_player.py
+python3 field_player.py
